requirements.txt
# Use generic, compatible versions
torch
torchvision
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
streamlit>=1.25.0
pydantic>=2.0.0
python-multipart>=0.0.6
pyyaml>=6.0
onnx>=1.14.0
onnxruntime>=1.15.0
pytest>=7.4.0
tqdm>=4.65.0
opencv-python>=4.8.0